open_bci_v3.py: Remove unnecessary bytes()
open_bci_v3.py seems to try to use bytes() to create a byte from a specified code.  However this is not what bytes() does; rather, it creates a string representation of a number (or a sequence of null bytes of a certain length in Python 3).  The resulting behavior is still OK since comparing the strings obtained this way will still work correctly; but it's a weird thing to do.

Another option would be to use '\xA0' and '\xC0' directly and not using struct.unpack() at all, but I think it's better to keep it this way for consistency with the rest of conversions.